Refactor tests of ad framework (#1007)
* MAINT: Move files to final destination.

* MAINT: Move files to final destination.

* MAINT: Move files to final destination.

* MAINT: test_ad.py content is moved to test_ad_array_forward_mode.py

* MAINT: test_grid_operators.py content is moved to test_ad_operator_tree.py

* MAINT: Conver test to pytest

* MAINT: Delete obsolete filetest_ad.py

* MAINT: Delete obsolete file test_grid_operators.py

* MAINT: Rename test_ad_array_forward_mode.py to test_forward_mode.py

* MAINT: Rename test_forward_ad_standard_functions.py to test_functions.py

* MAINT: Rename test_ad_operator_tree.py to test_operators.py

* MAINT: test_equation_system_evaluation.py to test_equation_system.py

* MAINT: refactor test_forward_mode.py

* MAINT: refactor test_functions.py

* MAINT: refactor test_operators.py

* MAINT: refactor test_equation_system.py

* MAINT: Move files to final destination.

* MAINT: test_ad.py content is moved to test_ad_array_forward_mode.py

* MAINT: test_grid_operators.py content is moved to test_ad_operator_tree.py

* MAINT: Conver test to pytest

* MAINT: Delete obsolete filetest_ad.py

* MAINT: Delete obsolete file test_grid_operators.py

* MAINT: Rename test_ad_array_forward_mode.py to test_forward_mode.py

* MAINT: Rename test_forward_ad_standard_functions.py to test_functions.py

* MAINT: Rename test_ad_operator_tree.py to test_operators.py

* MAINT: test_equation_system_evaluation.py to test_equation_system.py

* MAINT: refactor test_forward_mode.py

* MAINT: refactor test_functions.py

* MAINT: refactor test_operators.py

* MAINT: refactor test_equation_system.py
